wbank: Added frontend
Added an frontend app which uses react and redux. For the communication with the server
is used the axios library as it have nice interceptor api.

For the structure of the project is used redux ducks which makes actions and reducers closer
and are easy for modification.

The implemented functionalities are:
 * creation of new accounts and listing the existing
 * deposit and withdraw
 * viewing history of transactions

Fixes #2
Fixes #3
Fixes #4
Fixes #5